卡品报价

+++ club_module/src/main/java/com/yijian/clubmodule/bean/CardInfo.kt
@@ -5,13 +5,13 @@ package com.yijian.clubmodule.bean
  * email：850716183@qq.com
  * time: 2018/3/17 19:29:17
  */
-data class CardInfo(val cardprodbaseId: String, val cardType: Integer,
-                    val amount: Integer, val validDay: Integer,
-                    val rechargeGivePercent: Integer, val cardName: String, val venusNames: String,
-                    val rechargeGiveStr: String, val validTime: Integer, val salePrice: Integer, val venueNameList: List<String>) : Comparable<CardInfo> {
+data class CardInfo(val cardprodbaseId: String, val cardType: Int,
+                    val amount: Int, val validDay: Int,
+                    val rechargeGivePercent: Int, val cardName: String, val venusNames: String,
+                    val rechargeGiveStr: String, val validTime: Int, val salePrice: Int, val venueNameList: List<String>) : Comparable<CardInfo> {
 
 
     override fun compareTo(other: CardInfo): Int {
-        return salePrice.toInt() - other.salePrice.toInt()
+        return salePrice - other.salePrice
     }
 }
